pointer implementation -- going to scrap this though
Initial tests showing it's not as performant as it needs to be. Makes
sense in hindsight, these objects aren't discarded often enough or when
they're old enough, for gc to be the perf issue. Replacing resolve()
string munging with pointer lookups is about equivalent. Time to use
plain old JS entry objects instead.