use array instead of linked list for children
Also, use readdir instead of opendir for the async readdir case.  This
yields *significant* improvement in readdir/walk performance, owing to
the fact that `for await` iteration is excessively slow in V8 right now.
Which, is really a shame, because it's such a nice coding pattern, so I
do hope they get around to optimizing it eventually.  But it's fairly
new-ish, and probably not in widespread usage yet, so it's not terribly
surprising that it hasn't gotten the first class treatment yet.

Anyway, in testing, this is faster up until around 2**16 directory
entries, at which point the time spent creating the array of short
strings starts to tip into the bad zone.

*Synchronous* iteration is plenty fast, actually slightly more
performant than doing the full readdirSync and walking the array, so
it's a slightly unusual case here where the synchronous version is
slightly more resilient against extreme file system edge cases.