Move guts over to Path class
This also adds a gc() call on the path.resolve and readdir benchmarks.

And in the process of doing this, it seems like the perf issue that was
making the async generator style readdir() slow has gone away?  At
least, I can't seem to reproduce it any more, but it definitely comes
back when the readdir() method is on the PathWalker class.  Could be
that just extracting it keeps the types more consisent?  Idk, but the
API is cuter, so it stays.

Now the PathWalker class almost exclusively returns strings, unless
{withFileTypes: true} is passed.  (With the exception of `lstat`, since
the entire point is to get a path entry info object.)

Next: make most of the Path internals private.