rename module
Turns out 'path-walker' is available on the registry, but likely only
because there's an abandoned 10 year old module called 'pathwalker', and
the name is too similar to be used.

Successfully published an empty 'path-scurry' 0.0.0-0 module, so I know that
one works.

And it kind of captures the nature of what this thing is doing anyway.
It's going really fast, hurrying along, not covering the same ground
more than once.  Kind of like it better than "walker", which kind of
sounds less exciting.